# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArmorX Arms & Ammunition Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PRODUCTS
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: kycStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}/role:
    put:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Update user role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/kyc:
    get:
      operationId: adminListKyc
      tags: [admin]
      summary: List all KYC submissions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: KYC list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KycDetail"
  /admin/kyc/{id}/approve:
    post:
      operationId: adminApproveKyc
      tags: [admin]
      summary: Approve KYC submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /admin/kyc/{id}/reject:
    post:
      operationId: adminRejectKyc
      tags: [admin]
      summary: Reject KYC submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /admin/licenses/{id}/approve:
    post:
      operationId: adminApproveLicense
      tags: [admin]
      summary: Approve license (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}/reject:
    post:
      operationId: adminRejectLicense
      tags: [admin]
      summary: Reject license (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/audit-logs:
    get:
      operationId: adminListAuditLogs
      tags: [admin]
      summary: List audit logs (admin)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        agreedToTerms:
          type: boolean
      required: [email, password, firstName, lastName, dateOfBirth, agreedToTerms]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [guest, verified_buyer, admin]
        kycStatus:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        isAgeVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, firstName, lastName, role, kycStatus, isAgeVerified, createdAt]

    # PRODUCT SCHEMAS
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required: [productId, productName, price, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, compliance_review, approved, processing, shipped, delivered, cancelled, rejected]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        total:
          type: number
        shippingAddress:
          type: string
        complianceNotes:
          type: string
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, status, items, subtotal, total, shippingAddress, createdAt, updatedAt]
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
...
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, compliance_review, approved, processing, shipped, delivered, cancelled, rejected]
        complianceNotes:
          type: string
          nullable: true
        trackingNumber:
          type: string
          nullable: true
      required: [status]

    # KYC SCHEMAS
    KycStatus:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
      required: [status]
    KycDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userFullName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        documentType:
          type: string
        documentNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: string
        submittedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
          nullable: true
      required: [id, userId, userEmail, userFullName, status, documentType, documentNumber, dateOfBirth, address, submittedAt]
...
    License:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        licenseType:
          type: string
          enum: [firearms_dealer, concealed_carry, hunting, collector]
        licenseNumber:
          type: string
        issuingAuthority:
          type: string
        issuedDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, licenseType, licenseNumber, issuingAuthority, issuedDate, expiryDate, status, createdAt]
...
    AdminUserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, totalPages]
    UpdateRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [guest, verified_buyer, admin]
      required: [role]
    RejectBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        userEmail:
          type: string
          nullable: true
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, action, resourceType, createdAt]
    AuditLogListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, totalPages]

    # DASHBOARD SCHEMAS
...